Fix QDir::removeRecursively so it doesn't follow symlinks to directories.
Critical bug... Good thing I had backups of my $HOME.

+++ src/corelib/io/qdir.cpp
@@ -1502,7 +1502,7 @@ bool QDir::removeRecursively()
         di.next();
         const QFileInfo& fi = di.fileInfo();
         bool ok;
-        if (fi.isDir())
+        if (fi.isDir() && !fi.isSymLink())
             ok = QDir(di.filePath()).removeRecursively(); // recursive
         else
             ok = QFile::remove(di.filePath());
